ajustes incompletos
