+++ src/server.ts
@@ -1,12 +1,19 @@
-import Fastify, { type FastifyReply, type FastifyRequest } from 'fastify'
-import userRoutes from './domain/users/user.route'
+import buildApp from './app'
 
-function buildServer() {
-	const server = Fastify()
+const server = buildApp()
 
-	server.register(userRoutes, { prefix: 'api/users' })
+async function main() {
+	try {
+		await server.listen({
+			host: '0.0.0.0',
+			port: process.env.PORT ? Number(process.env.PORT) : 3333,
+		})
 
-	return server
+		console.log('HTTP Server started - http://localhost:3000')
+	} catch (e) {
+		console.log(`HTTP Server stoped ERROR: ${e}`)
+		process.exit(1)
+	}
 }
 
-export default buildServer
+main()
